cleanup repro
