+++ Xamarin.Forms.Controls/Xamarin.Forms.Controls.csproj
@@ -84,6 +84,7 @@
   <ItemGroup>
     <Compile Include="App.cs" />
     <Compile Include="AppLifeCycle.cs" />
+    <Compile Include="Bugzilla44596SplashPage.cs" />
     <Compile Include="ControlGalleryPages\CellForceUpdateSizeGalleryPage.cs" />
     <Compile Include="ControlGalleryPages\LayoutAddPerformance.xaml.cs">
       <DependentUpon>LayoutAddPerformance.xaml</DependentUpon>
